fix(pubsub): Vite build issues (#10353)
